% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geofcm.R
\name{geofcm}
\alias{geofcm}
\title{GEOFCM based on Yoshida et al. (2018)}
\usage{
geofcm(x, coord, C, m = 1.5, q = 1.5, r = 5, scaling = TRUE,
  iter.max = 100L, th = 1e-05)
}
\arguments{
\item{x}{matrix or data.frame containing observed data other than coordinates}

\item{coord}{matrix or data.frame containing coordinate date}

\item{C}{number of clusters}

\item{m}{Controls fuzziness based on (default = 1.5)}

\item{q}{Controls fuzziness based on distance (default = 1.5)}

\item{r}{Distance in km to find neighbors of each data point. (default = 5)}

\item{scaling}{TRUE or FALSE to scale x (default = TRUE)}

\item{maximum}{number of iterations (default = 100L).}

\item{threshold}{to determine convergence (default = 1e-5).}
}
\description{
GEOFCM based on Yoshida et al. (2018)
}
