% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geofcm.R
\name{plot.GEOFCM}
\alias{plot.GEOFCM}
\title{plot for GEOFCM}
\usage{
\method{plot}{GEOFCM}(x, ...)
}
\arguments{
\item{x}{GEOFCM class object}

\item{...}{other arguments passed to print}

\item{n}{number of rows to view matrix of membership degrees}
}
\description{
plot for GEOFCM
}
