% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geofcm.R
\name{print.GEOFCM}
\alias{print.GEOFCM}
\title{print for GEOFCM}
\usage{
\method{print}{GEOFCM}(x, n = 5, plot = TRUE, ...)
}
\arguments{
\item{x}{GEOFCM class object}

\item{n}{number of rows to view matrix of membership degrees}

\item{plot}{TRUE or FALSE to plot the result}

\item{...}{other arguments passed to print}
}
\description{
print for GEOFCM
}
